+++ src/Module.php
@@ -15,7 +15,15 @@
 namespace blackcube\core;
 
 use blackcube\core\commands\InitController;
+use blackcube\core\components\Plugins;
+use blackcube\core\components\PluginsHandler;
+use blackcube\core\helpers\PluginHelper;
+use blackcube\core\interfaces\PluginInterface;
+use blackcube\core\interfaces\PluginManagerRoutableInterface;
+use blackcube\core\interfaces\PluginServiceInterface;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use blackcube\core\models\Parameter;
+use blackcube\core\models\Plugin;
 use blackcube\core\web\UrlRule;
 use blackcube\core\web\UrlMapper;
 use creocoder\flysystem\Filesystem;
@@ -106,6 +114,11 @@ class Module extends BaseModule implements BootstrapInterface
     public $allowedParameterDomains = [];
 
     /**
+     * @var array plugins definitions
+     */
+    public $plugins = [];
+
+    /**
      * @var CacheInterface|array|string|null
      */
     public $cache;
@@ -132,6 +145,7 @@ class Module extends BaseModule implements BootstrapInterface
 
     }
 
+
     /**
      * @inheritdoc
      */
@@ -145,8 +159,30 @@ class Module extends BaseModule implements BootstrapInterface
         if ($app instanceof WebApplication) {
             $this->bootstrapWeb($app);
         }
+        $this->registerPlugins($app);
     }
 
+    public function registerPlugins($app)
+    {
+        if (Yii::$container->hasSingleton(PluginsHandlerInterface::class) === false) {
+            Yii::$container->setSingleton(PluginsHandlerInterface::class, PluginsHandler::class);
+        }
+        if ($app instanceof WebApplication) {
+            $pluginHandlerUrlManager = Yii::createObject(PluginsHandlerInterface::class);
+            $app->getUrlManager()->addRules([
+                $pluginHandlerUrlManager
+            ], true);
+            foreach($pluginHandlerUrlManager->getActivePluginManagers() as $pluginManager) {
+                if ($pluginManager instanceof PluginManagerRoutableInterface) {
+                    foreach($pluginManager->getControllerMap() as $id => $controllerMap) {
+                        if (isset($app->controllerMap[$id]) === false) {
+                            $app->controllerMap[$id] = $controllerMap;
+                        }
+                    }
+                }
+            }
+        }
+    }
     /**
      * Init console stuff
      *